Chat user initialization
