+++ store/store.ts
@@ -1,5 +1,6 @@
 import { TradingHubStateType } from '@/components/Market/TradingHub/TradingHub';
 import { EnrichedMarketType } from '@/types/marketTypes';
+import { PushAPI } from '@pushprotocol/restapi';
 import { atom } from 'jotai';
 
 /* MARKETS ATOM */
@@ -25,3 +26,7 @@ export const chosenInterlocutorAtom = atom<string>(
 /* CHART INTERVAL */
 
 export const chartIntervalAtom = atom<'15M' | '1H'>('1H');
+
+/* CHAT INITIALIZED USER */
+
+export const chatUserAtom = atom<PushAPI | undefined>(undefined);
